Tidied up some files
