+++ inc/ai_aircraft.h
@@ -27,7 +27,7 @@ typedef struct
 
 void ai_aircraft_init(AiAircraft *aircraft, int num_aircraft);
 void ai_aircraft_update_pose(AiAircraft *aircraft, int num_aircraft, int frame_counter);
-void ai_aircraft_create(AiAircraft *dest, const AiAircraft *src, int count);
+void ai_aircraft_copy(AiAircraft *dest, const AiAircraft *src, int count);
 void ai_aircraft_move(AiAircraft *aircraft, Heading heading);
 void ai_aircraft_update_status(AiAircraft * crafts, int num_aircraft);
 void normalize_vector(double *dx, double *dy, double *dz);
