+++ src/app.c
@@ -14,23 +14,23 @@
  * sections of the MSLA applicable to Source Code.
  *
  ******************************************************************************/
-#include "sl_sleeptimer.h"
-#include "game.h"
-#include "sl_sleeptimer.h"
-#include "game.h"
-#include "em_gpio.h"
-#include "debug.h"
 
-#ifndef FPS
-#define FPS 60
-#endif
+#include "sl_sleeptimer.h"
+#include "game.h"
+#include "app.h"
+#include "display.h"
 
 sl_sleeptimer_timer_handle_t frame_timer;
 bool frame_ready = false;
 uint32_t frame_counter = 0;
 uint32_t game_active;
 
-static void on_frame_timeout(sl_sleeptimer_timer_handle_t *handle, void *data);
+static void on_frame_timeout(sl_sleeptimer_timer_handle_t *handle, void *data)
+{
+  (void)handle;
+  (void)data;
+  frame_ready = true;
+}
 
 void frame_timer_init()
 {
@@ -45,6 +45,7 @@ void app_init()
   game_init();
   game_active = 0;
   game_process_action(frame_counter, &game_active);
+  display_set_string("WAITING TO START MICROHARD FLIGHT SIMULATOR ");
 }
 
 void app_process_action(void)
@@ -54,9 +55,6 @@ void app_process_action(void)
   {
     frame_counter++;
     frame_ready = false;
-//    debug_print_float((float) game_active);
-//    debug_printf("\n");
-
     if (game_active)
       {
         game_process_action(frame_counter, &game_active);
@@ -68,9 +66,4 @@ void app_process_action(void)
   }
 }
 
-static void on_frame_timeout(sl_sleeptimer_timer_handle_t *handle, void *data)
-{
-  (void)handle;
-  (void)data;
-  frame_ready = true;
-}
+
